Moved the cookbook out path to $CHARM_DIR
This allows the isolation of cookbooks outside of the hooks path, keeping
concerns separate. This is a precursor to more functinality landing with regard
to berkshelf cookbook management, and additional chef patterns. The lions share
of the bootstrap script will go away soon(ish) and land in a charm-helper which
will take care of installing the ruby dependencies, setting up chef-solo and
prepping the environment. Before that can happen, the charm template needs
to undergo some permutations.