Implement charm-proof for storage
Added colander dep for schema validation. Already packaged for apt
so this shouldn't add any packaging burden down the road.

There were a handful of +x files in the tests/ dir, which was
(curiously) causing nosetests to not find those tests. After setting
-x, nose detected about 50 additional tests that weren't being run
before. Many of those (in test_juju_test.py) were failing because
the cwd was not being set properly to allow the Conductor to find
the tests. So that's what all those updates are for...

The change to record.sh is to ensure the tests run successfully
after a clean repo checkout (they failed for me originally).