Purge duplicate wheels for module when overriding
When the --wheelhouse-overrides option is used, duplicate wheels can
be produced in the resulting wheelhouse if the version pulled in by
the override wheelhouse is different from the version pulled in
during the regular build.

With this patch when the override wheelhouse is being processed it
removes any existing wheelhouses for the module being added. Since
the override wheelhouse is processed last the modules specified in
it are always the copies from the override.

Looking at wheel package format *1 hypens are replaced with
underscores in the distribution name and then hypens are used as
delimiters in the wheel file name *2:
{distribution}-{version}(-{build tag})?-{python tag}-{abi tag}-{platform
tag}.whl.
Given that I think deleting wheels that match {distribution}-* is
safe.

*1 https://www.python.org/dev/peps/pep-0427/#escaping-and-unicode
*2 https://www.python.org/dev/peps/pep-0427/#file-name-convention

closes #381