Support charms.reactive layer options
Allow for the reactive library to have its own options section even
though it's not a layer by itself.  We might want to support a more
generic way of handling this, so that arbitrary libraries can defined
their own sections, perhaps via entry points, but this is a quick fix
for the reactive library itself.