Restore historical behavior for --output-dir
Previous to #478, the behavior of `--output-dir` matched that of
`$JUJU_REPOSITORY` in that it specified a directory which would have a
`builds` or `{series}` directory created under it, which would then have
the final charm output dir created under that.  This got changed and
broke the charm builds for OpenStack, so this restores the previous
behavior of `--output-dir` while retaining the new semantics for
`--build-dir`.