Add promulgate and unpromulgate commands
This adds promulgate and unpromulgate commands to the charm snap.  These
commands do some basic integrity checking to ensure that things don't
get promulgated into a bad state, including:

  * Ensuring that there isn't already a charm promulgated with that name
  * Ensuring that the charm is released to stable
  * Ensuring that the stable branch is readable by "everyone"
  * After promulgation, ensuring that the owner can release updates to
    their charm

See https://discourse.jujucharms.com/t/promulgate-a-charm-for-charmers-team/1206

This also updates the snap build to use core18, which is required for
bhttp.