Fix and improve charm-layers
The `charm layers` command was broken with the addition of revisions to
the manifest.  This fixes that, and also adds some additional
functionality, such as the ability to annotate the source layer when
colors aren't available, and a larger number of colors to choose from so
that charms with many layers won't run out quite as soon.