Process wheelhouse.txt wholistically rather than per-layer
Processing the wheelhouse.txt file from each layer in isolation means
that pip can't resolve conflicts, duplicates, or dependency version
pinning which spans across layers.  This can lead to duplicate versions
of libraries being installed which violate constraints to avoid certain
upstream bugs.

This came up in both of:

* https://bugs.launchpad.net/charm-flannel/+bug/1884598
* https://bugs.launchpad.net/charm-octavia/+bug/1884164

And in both cases trying to resolve it on a per-layer or per-charm basis
results in a whack-a-mole process of catching all of the layers which
indirectly reintroduce the unpinned version of a dependency pinned in a
base layer.