Raise more useful BuildError on missing pkg name
When a URL style requirement is used in a wheelhouse without a package
name, we can't determine whether it should override something from a
previous layer or not. Previously, this raised a cryptic TypeError due
to None being used in a regexp match. This changes it to raise a
BuildError with useful info for correcting the problem.