Add --branch option to pull-source
Adds the ability to specify a branch to switch to when fetching the
source of a charm or layer.  This is particularly useful for layer
entries which use the `subdir` field, since that can lose the SCM
metadata and prevent post-pull switching.

Fixes #580
Fixes [lp:1893081](https://bugs.launchpad.net/charmed-kubernetes-testing/+bug/1893081)