Add reproducible charm build feature (#585)
* Add reproducible charm build feature

This feature adds some command line options to allow the generation of
the build.lock file in the top layer.  This can then be used to rebuild
the charm with the exact same set of layers and python modules. Also the
ability to track a specific branch is also enabled.

* Update reproducible charms doc with suggestions