Fix hard-coded WheelhouseTactic in builder.py
When doing either the reproducible charms or the wheelhouse overrides,
the builder used to use the WheelhouseTactic directly.  This breaks any
code that might override the WheelhouseTactic.  Instead, if an existing
wheelhouse tactic is in use, use that one's class; if it's not, then
default back to the base WheelhouseTactic.  May help with bug #585