Fix reproducible charms issues (#598)
* Fix issues with recent reproducible charms feature

The issues fixed in this are:

 * --ignore-lock-file didn't fully ignore the lock file.
 * rebuilding resulted in missing libraries due to purging the wheels.
   Purging was not needed.
 * The wrong fetcher classes were used when rebuilding from lock-files.
   This had the side-effect of not being able to handle multi-layer git
   repositories.  This has been resolved.

* Fix hard-coded WheelhouseTactic in builder.py

When doing either the reproducible charms or the wheelhouse overrides,
the builder used to use the WheelhouseTactic directly.  This breaks any
code that might override the WheelhouseTactic.  Instead, if an existing
wheelhouse tactic is in use, use that one's class; if it's not, then
default back to the base WheelhouseTactic.  May help with bug #585

* Fix test problems with missing charm_file from Builder()

It's a bit odd how these passed in the past; adding it as a default to
the class resolves the unit test issues.

* Add build.lock to ignores so it doesn't end up in the built charm