snap: Replace bhttp part with stage-snap
The approach used in the snapcraft `godeps` plugin is deprecated
and no longer works.

Replace it with using the `bhttp` snap as a stage-snap.

Note that this part of charm tools is probably not useful anymore
as the old charm store is shut down, but let's get the snap
building again.