snap: Update base to core22
The `git` binary provided with a core18 base is too old to
work with modern versions of pip.

Update paths to match the new base.

Override LD_LIBRARY_PATH because the charm tool must be able
to execute binaries both from within the snap and from the
host operating system.

Drop python plugin as that is not usable with a core22 classic
snap.  Add manual patchelf part as the nil plugin does not do
this automatically.

Drop -dev stage packages as those pull in the libc6 package
that contain absolute symlinks pointing at /, which is not
allowed in a core22 classic snap.

Fixes: #625