snap: Ensure build succeeds on non-x86 architectures
Update snap recipe so that building charm-tools itself succeeds
on non-x86 architectures.

With reproducibility and maintainability of the snap in mind,
make use of Python dependencies from Ubuntu distribution
packages wherever possible.

For any dependencies not available in the Ubuntu distribution, we
request a build from source as this will ensure that build-package
requirements are uncovered at a early stage in the development
process, preferably already in the gate check.

Drop the PYTHON related build-environment variables as these would
mandate installing all build dependencies in the snap, which is
not what we want.  Instead reference these variables when needed
in the build recipe.