Prune GIT environment variables when using Python from buildenv
At present we call out to Python provided by the build environment
to create a correct wheelhouse.

However, we are not pruning GIT related environment variables set
by the snap.  This in turn makes git clone fail when the `git`
binary from the build environment attempts to use git resources in
the snap, which may be of a different version.

Fixes #632