Add integration test
Colocate all jobs in a single file in order to make use of job
dependencies.

Inject a fake git tag in the snap build job so that valid version
information is provided in the build.  Attempts to adjust
fetch-depth or direct calls to `git fetch --tags` was
unsuccessful.  The charmcraft reactive plugin validates this
information later in the integration test.

Add integration test that builds a minimal reactive charm using
charmcraft and the reactive plugin.