Add option to upgrade core deps in build venv
Some packages may require newer versions of core dependencies for
successful execution of the `pip download` or `pip wheel`
commands. Most notable example is Jinja2 and MarkupSafe when
installed on Focal ref pallets/jinja#1496.

At this stage of the process the input from the charm
wheelhouse.txt is not sufficient.

Add an option to upgrade pip and setuptools in the build venv to
deal with these situations.

Fixes: #646
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>