Pin pip<23.1 and setuptools<67 to allow PEP-440 non compliance
PEP-440 is more strict about requirement lines.  As an example:

    pytz>dev

is non-compliant, but is in 5.2.4 of kombu's requirements.  This breaks
building wheelhouses from source (such as the octavia charm).  This
patch pins pip and setuptools to the latest versions that will defintely
still allow PEP-440 non-compliant packages to install.

The option --upgrade-buildvenv-core-deps can be used to override this
and will install the latest versions of pip and setuptools available.

Fixes-Bug: #652