Upgrade setuptools
setuptools-scm depends on setuptools>=61, while the one installed in the
building environment is setuptools-59.x which comes from the deb package
available in Ubuntu-22.04.

This change upgrades setuptools to allow the building to succeed. Also
uses an absolute path when running `vergit` to fix a "command not found"
error.

Fixes #666