Migrate pkg_resources to importlib.resources
The pkg_resources has been deprecated and produces a warning, this is
preventing charmcraft from reading the version correctly, since the json
output gets mangled due to the warning.

This change introduces the use of importlib.resources and supresses the
warning for the `vergit` module which we don't have control of at the
moment.